fix: Hash() now uses the 32 bit version of FNV-1a

+++ include/parsing.h
@@ -8,7 +8,7 @@ typedef struct {
 
 int GetSlice(bufState* buf_state, Slice* slice, const char* delim, const int delim_len);
 
-unsigned long Hash(void* input, size_t size);
+unsigned int Hash(void* input, size_t size);
